requirements.txt
# Core API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Scientific Computing
numpy==1.26.3
scipy==1.12.0

# HTTP & Web
requests==2.32.3
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Utils
python-dotenv==1.0.0

# Optional: Streamlit (for legacy)
streamlit==1.39.0
